refactor : 뷰모델 생성 후 코드 내용 이전
