+++ app/src/main/res/layout/fragment_game_start.xml
@@ -6,7 +6,7 @@
         <import type="com.ursssu.unscramble.R" />
         <variable
             name="viewModel"
-            type="com.ursssu.unscramble.presentation.GameStartFragment" />
+            type="com.ursssu.unscramble.presentation.GameStartViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -24,7 +24,8 @@
             android:gravity="center"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toTopOf="parent"
+            app:text = "@{viewModel.gameStartTimerText}"/>
 
         <com.yourssu.design.system.atom.Text
             android:id="@+id/text_game_start_word"
@@ -33,15 +34,17 @@
             android:layout_marginTop="70dp"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="@id/text_game_start_timer" />
+            app:layout_constraintTop_toTopOf="@id/text_game_start_timer"
+            app:text = "@{viewModel.gameStartWordText}"
+            />
 
         <com.yourssu.design.system.atom.SimpleTextField
             android:id="@+id/textfield_game_start"
             android:layout_width="match_parent"
             android:layout_height="100dp"
-            android:layout_marginLeft="20dp"
+            android:layout_marginHorizontal="20dp"
             android:layout_marginTop="100dp"
-            android:layout_marginRight="20dp"
+            app:helperLabelText="@{viewModel.textFieldHelperLabelText}"
             app:layout_constraintTop_toTopOf="@id/text_game_start_word"
             />
